requirements.txt
Flask==2.3.3
opencv-python==4.8.1.78
Pillow==10.0.1
moviepy==1.0.3
numpy==1.24.3
PyPDF2==3.0.1
python-docx==0.8.11
gTTS==2.5.1
pydub==0.25.1
